--- apps/client/src/app/best/layout.tsx
@@ -1,9 +1,0 @@
-import React from "react"
-
-export default function layout({
-	children,
-}: {
-	children: Readonly<React.ReactNode>
-}) {
-	return <>{children}</>
-}
