+++ apps/client/src/components/best/organism/BestTop5.tsx
@@ -1,22 +1,22 @@
 import React from "react"
-import type { BestCreatorDataTypes } from "@/types/best/bestTypes"
-import BestTop5Card from "../molecule/BestTop5Card"
+import type { RenderedRankingItemTypes } from "@/types/best/bestTypes"
+import BestTop5Card from "@/components/best/molecule/BestTop5Card"
 
 interface BestListProps<T> {
-	data: T[]
+	data: T
 }
-function BestTop5({ data }: BestListProps<BestCreatorDataTypes>) {
+function BestTop5({ data }: BestListProps<RenderedRankingItemTypes[]>) {
 	const top1Data = data[0]
 	const restData = data.slice(1, 5)
 	return (
 		<div className="container mx-auto p-8">
 			<div className="grid grid-cols-1 gap-8">
 				<div className="flex justify-center">
-					<BestTop5Card key={top1Data.id} {...top1Data} />
+					<BestTop5Card key={top1Data.memberUuid} {...top1Data} />
 				</div>
 				<div className="grid grid-cols-1 gap-4 md:grid-cols-2 md:gap-6 lg:grid-cols-4 lg:gap-8">
 					{restData.map((creator) => (
-						<BestTop5Card key={creator.id} {...creator} />
+						<BestTop5Card key={creator.memberUuid} {...creator} />
 					))}
 				</div>
 			</div>
@@ -25,3 +25,4 @@ function BestTop5({ data }: BestListProps<BestCreatorDataTypes>) {
 }
 
 export default BestTop5
+
